requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
crewai==0.11.0 